+++ mps_youtube/main.py
@@ -58,9 +58,12 @@ from urllib.parse import urlencode
 
 import pafy
 
-from . import terminalsize, g
+from . import terminalsize, g, c
 from .playlist import Playlist, Video
 from .paths import get_default_ddir, get_config_dir
+from .config import Config, known_player_set
+from .util import has_exefile, get_mpv_version, dbg
+from .util import xenc, xprint, mswinfn, set_window_title, clear_screen
 
 try:
     # pylint: disable=F0401
@@ -108,94 +111,6 @@ def getxy():
         y = max_results + 4
 
     return XYTuple(x, y, max_results)
-
-
-def utf8_replace(txt):
-    """ Replace unsupported characters in unicode string, returns unicode. """
-    sse = sys.stdout.encoding
-    txt = txt.encode(sse, "replace").decode(sse)
-    return txt
-
-
-def xenc(stuff):
-    """ Replace unsupported characters. """
-    if sys.stdout.isatty():
-        return utf8_replace(stuff) if not_utf8_environment else stuff
-
-    else:
-        return stuff.encode("utf8", errors="replace")
-
-
-def xprint(stuff, end=None):
-    """ Compatible print. """
-    print(xenc(stuff), end=end)
-
-
-def mswinfn(filename):
-    """ Fix filename for Windows. """
-    if mswin:
-        filename = utf8_replace(filename) if not_utf8_environment else filename
-        allowed = re.compile(r'[^\\/?*$\'"%&:<>|]')
-        filename = "".join(x if allowed.match(x) else "_" for x in filename)
-
-    return filename
-
-
-def set_window_title(title):
-    """ Set terminal window title. """
-    if mswin:
-        os.system(xenc("title " + title))
-    else:
-        sys.stdout.write(xenc('\x1b]2;' + title + '\x07'))
-
-
-def clear_screen():
-    """Clear all text from screen."""
-    if g.no_clear_screen:
-        xprint('--\n')
-    elif mswin:
-        os.system('cls')
-    elif has_exefile('tput'):
-        subprocess.call(['tput', 'reset'])
-    else:
-        xprint('\n' * 200)
-
-
-def get_mpv_version(exename):
-    """ Get version of mpv as 3-tuple. """
-    o = subprocess.check_output([exename, "--version"]).decode()
-    re_ver = re.compile(r"%s (\d+)\.(\d+)\.(\d+)" % exename)
-
-    for line in o.split("\n"):
-        m = re_ver.match(line)
-
-        if m:
-            v = tuple(map(int, m.groups()))
-            dbg("%s version %s.%s.%s detected", exename, *v)
-            return v
-
-    dbg("%sFailed to detect mpv version%s", c.r, c.w)
-    return -1, 0, 0
-
-
-def has_exefile(filename):
-    """ Check whether file exists in path and is executable.
-
-    Return path to file or False if not found
-    """
-    paths = [os.getcwd()] + os.environ.get("PATH", '').split(os.pathsep)
-    paths = [i for i in paths if i]
-    dbg("searching path for %s", filename)
-
-    for path in paths:
-        exepath = os.path.join(path, filename)
-
-        if os.path.isfile(exepath):
-            if os.access(exepath, os.X_OK):
-                dbg("found at %s", exepath)
-                return exepath
-
-    return False
 
 
 def get_content_length(url, preloading=False):
@@ -360,322 +275,6 @@ def get_size(ytid, url, preloading=False):
         dbg("%s%s - content-length: %s%s", c.y, prefix, stream['size'], c.w)
 
     return stream['size']
-
-
-class ConfigItem(object):
-
-    """ A configuration item. """
-
-    def __init__(self, name, value, minval=None, maxval=None, check_fn=None,
-            require_known_player=False, allowed_values=None):
-        """ If specified, the check_fn should return a dict.
-
-        {valid: bool, message: success/fail mesage, value: value to set}
-
-        """
-        self.default = self.value = value
-        self.name = name
-        self.type = type(value)
-        self.maxval, self.minval = maxval, minval
-        self.check_fn = check_fn
-        self.require_known_player = require_known_player
-        self.allowed_values = []
-        if allowed_values:
-            self.allowed_values = allowed_values
-
-    @property
-    def get(self):
-        """ Return value. """
-        return self.value
-
-    @property
-    def display(self):
-        """ Return value in a format suitable for display. """
-        retval = self.value
-
-        if self.name == "max_res":
-            retval = str(retval) + "p"
-
-        if self.name == "encoder":
-            retval = str(retval) + " [%s]" % (str(g.encoders[retval]['name']))
-
-        return retval
-
-    def set(self, value):
-        """ Set value with checks. """
-        # note: fail_msg should contain %s %s for self.name, value
-        #       success_msg should not
-        # pylint: disable=R0912
-        # too many branches
-
-        success_msg = fail_msg = ""
-        value = value.strip()
-        value_orig = value
-
-        # handle known player not set
-
-        if self.allowed_values and value not in self.allowed_values:
-            fail_msg = "%s must be one of * - not %s"
-            allowed_values = self.allowed_values.copy()
-            if '' in allowed_values:
-                allowed_values[allowed_values.index('')] = "<nothing>"
-            fail_msg = fail_msg.replace("*", ", ".join(allowed_values))
-
-        if self.require_known_player and not known_player_set():
-            fail_msg = "%s requires mpv or mplayer, can't set to %s"
-
-        # handle true / false values
-
-        elif self.type == bool:
-
-            if value.upper() in "0 OFF NO DISABLED FALSE".split():
-                value = False
-                success_msg = "%s set to False" % c.c("g", self.name)
-
-            elif value.upper() in "1 ON YES ENABLED TRUE".split():
-                value = True
-                success_msg = "%s set to True" % c.c("g", self.name)
-
-            else:
-                fail_msg = "%s requires True/False, got %s"
-
-        # handle int values
-
-        elif self.type == int:
-
-            if not value.isdigit():
-                fail_msg = "%s requires a number, got %s"
-
-            else:
-                value = int(value)
-
-                if self.maxval and self.minval:
-
-                    if not self.minval <= value <= self.maxval:
-                        m = " must be between %s and %s, got "
-                        m = m % (self.minval, self.maxval)
-                        fail_msg = "%s" + m + "%s"
-
-                if not fail_msg:
-                    dispval = value or "None"
-                    success_msg = "%s set to %s" % (c.c("g", self.name),
-                                                    dispval)
-
-        # handle space separated list
-
-        elif self.type == list:
-            success_msg = "%s set to %s" % (c.c("g", self.name), value)
-            value = value.split()
-
-        # handle string values
-
-        elif self.type == str:
-            dispval = value or "None"
-            success_msg = "%s set to %s" % (c.c("g", self.name),
-                                            c.c("g", dispval))
-
-        # handle failure
-
-        if fail_msg:
-            failed_val = value_orig.strip() or "<nothing>"
-            colvals = c.y + self.name + c.w, c.y + failed_val + c.w
-            return fail_msg % colvals
-
-        elif self.check_fn:
-            checked = self.check_fn(value)
-            value = checked.get("value") or value
-
-            if checked['valid']:
-                value = checked.get("value", value)
-                self.value = value
-                Config.save()
-                return checked.get("message", success_msg)
-
-            else:
-                return checked.get('message', fail_msg)
-
-        elif success_msg:
-            self.value = value
-            Config.save()
-            return success_msg
-
-
-def check_console_width(val):
-    """ Show ruler to check console width. """
-    valid = True
-    message = "-" * val + "\n"
-    message += "console_width set to %s, try a lower value if above line ove"\
-        "rlaps" % val
-    return dict(valid=valid, message=message)
-
-
-def check_api_key(key):
-    """ Validate an API key by calling an API endpoint with no quota cost """
-    url = "https://www.googleapis.com/youtube/v3/i18nLanguages"
-    query = {"part": "snippet", "fields": "items/id", "key": key}
-    try:
-        urlopen(url + "?" + urlencode(query)).read()
-        message = "The key, '" + key + "' will now be used for API requests."
-
-        # Make pafy use the same api key
-        pafy.set_api_key(Config.API_KEY.get)
-
-        return dict(valid=True, message=message)
-    except HTTPError:
-        message = "Invalid key or quota exceeded, '" + key + "'"
-        return dict(valid=False, message=message)
-
-
-def check_ddir(d):
-    """ Check whether dir is a valid directory. """
-    expanded = os.path.expanduser(d)
-    if os.path.isdir(expanded):
-        message = "Downloads will be saved to " + c.y + d + c.w
-        return dict(valid=True, message=message, value=expanded)
-
-    else:
-        message = "Not a valid directory: " + c.r + d + c.w
-        return dict(valid=False, message=message)
-
-
-def check_win_pos(pos):
-    """ Check window position input. """
-    if not pos.strip():
-        return dict(valid=True, message="Window position not set (default)")
-
-    pos = pos.lower()
-    reg = r"(TOP|BOTTOM).?(LEFT|RIGHT)"
-
-    if not re.match(reg, pos, re.I):
-        msg = "Try something like top-left or bottom-right (or default)"
-        return dict(valid=False, message=msg)
-
-    else:
-        p = re.match(reg, pos, re.I).groups()
-        p = "%s-%s" % p
-        msg = "Window position set to %s" % p
-        return dict(valid=True, message=msg, value=p)
-
-
-def check_win_size(size):
-    """ Check window size input. """
-    if not size.strip():
-        return dict(valid=True, message="Window size not set (default)")
-
-    size = size.lower()
-    reg = r"\d{1,4}x\d{1,4}"
-
-    if not re.match(reg, size, re.I):
-        msg = "Try something like 720x480"
-        return dict(valid=False, message=msg)
-
-    else:
-        return dict(valid=True, value=size)
-
-
-def check_encoder(option):
-    """ Check encoder value is acceptable. """
-    encs = g.encoders
-
-    if option >= len(encs):
-        message = "%s%s%s is too high, type %sencoders%s to see valid values"
-        message = message % (c.y, option, c.w, c.g, c.w)
-        return dict(valid=False, message=message)
-
-    else:
-        message = "Encoder set to %s%s%s"
-        message = message % (c.y, encs[option]['name'], c.w)
-        return dict(valid=True, message=message)
-
-
-def check_player(player):
-    """ Check player exefile exists and get mpv version. """
-    if has_exefile(player):
-
-        if "mpv" in player:
-            g.mpv_version = get_mpv_version(player)
-            version = "%s.%s.%s" % g.mpv_version
-            fmt = c.g, c.w, c.g, c.w, version
-            msg = "%splayer%s set to %smpv%s (version %s)" % fmt
-            return dict(valid=True, message=msg, value=player)
-
-        else:
-            msg = "%splayer%s set to %s%s%s" % (c.g, c.w, c.g, player, c.w)
-            return dict(valid=True, message=msg, value=player)
-
-    else:
-        if mswin and not player.endswith(".exe"):
-            return check_player(player + ".exe")
-
-        else:
-            msg = "Player application %s%s%s not found" % (c.r, player, c.w)
-            return dict(valid=False, message=msg)
-
-
-class _Config(object):
-
-    """ Holds various configuration values. """
-
-    _configitems = collections.OrderedDict((
-            ("ORDER", ConfigItem("order", "relevance",
-                allowed_values="relevance date views rating".split())),
-            ("USER_ORDER", ConfigItem("user_order", "", allowed_values =
-                [""] + "relevance date views rating".split())),
-            ("MAX_RESULTS", ConfigItem("max_results", 19,
-                maxval=50, minval=1)),
-            ("CONSOLE_WIDTH", ConfigItem("console_width", 80, minval=70,
-                maxval=880, check_fn=check_console_width)),
-            ("MAX_RES", ConfigItem("max_res", 2160, minval=192, maxval=2160)),
-            ("PLAYER", ConfigItem("player", "mplayer" + ".exe" * mswin,
-                check_fn=check_player)),
-            ("PLAYERARGS", ConfigItem("playerargs", "")),
-            ("ENCODER", ConfigItem("encoder", 0, minval=0,
-                check_fn=check_encoder)),
-            ("NOTIFIER", ConfigItem("notifier", "")),
-            ("CHECKUPDATE", ConfigItem("checkupdate", True)),
-            ("SHOW_MPLAYER_KEYS", ConfigItem("show_mplayer_keys", True,
-                require_known_player=True)),
-            ("FULLSCREEN", ConfigItem("fullscreen", False,
-                require_known_player=True)),
-            ("SHOW_STATUS", ConfigItem("show_status", True)),
-            ("COLUMNS", ConfigItem("columns", "")),
-            ("DDIR", ConfigItem("ddir", get_default_ddir(),
-                check_fn=check_ddir)),
-            ("OVERWRITE", ConfigItem("overwrite", True)),
-            ("SHOW_VIDEO", ConfigItem("show_video", False)),
-            ("SEARCH_MUSIC", ConfigItem("search_music", True)),
-            ("WINDOW_POS", ConfigItem("window_pos", "", check_fn=check_win_pos,
-                require_known_player=True)),
-            ("WINDOW_SIZE", ConfigItem("window_size", "",
-                check_fn=check_win_size, require_known_player=True)),
-            ("DOWNLOAD_COMMAND", ConfigItem("download_command", '')),
-            ("API_KEY", ConfigItem("api_key",
-                "AIzaSyCIM4EzNqi1in22f4Z3Ru3iYvLaY8tc3bo",
-                check_fn=check_api_key))
-        ))
-
-    def __getitem__(self, key):
-        return self._configitems[key]
-
-    def __getattr__(self, name):
-        try:
-            return self._configitems[name]
-        except KeyError:
-            raise AttributeError
-
-    def __iter__(self):
-        return iter(self._configitems.keys())
-
-    def save(self):
-        """ Save current config to file. """
-        config = {setting: self[setting].value for setting in self}
-
-        with open(g.CFFILE, "wb") as cf:
-            pickle.dump(config, cf, protocol=2)
-
-        dbg(c.p + "Saved config: " + g.CFFILE + c.w)
-
-Config = _Config()
 
 
 def get_version_info():
@@ -962,21 +561,6 @@ def init_readline():
             dbg(c.g + "Read history file" + c.w)
 
 
-def known_player_set():
-    """ Return true if the set player is known. """
-    for allowed_player in g.playerargs_defaults:
-        regex = r'(?:\b%s($|\.[a-zA-Z0-9]+$))' % re.escape(allowed_player)
-        match = re.search(regex, Config.PLAYER.get)
-
-        if mswin:
-            match = re.search(regex, Config.PLAYER.get, re.IGNORECASE)
-
-        if match:
-            return allowed_player
-
-    return None
-
-
 def showconfig(_):
     """ Dump config data. """
     width = getxy().width
@@ -1045,33 +629,6 @@ def import_config():
 
             Config.PLAYERARGS.value = " ".join(Config.PLAYERARGS.get)
             Config.save()
-
-
-class c(object):
-
-    """ Class for holding colour code values. """
-
-    if mswin and has_colorama:
-        white = Style.RESET_ALL
-        ul = Style.DIM + Fore.YELLOW
-        red, green, yellow = Fore.RED, Fore.GREEN, Fore.YELLOW
-        blue, pink = Fore.CYAN, Fore.MAGENTA
-
-    elif mswin:
-        ul = red = green = yellow = blue = pink = white = ""
-
-    else:
-        white = "\x1b[%sm" % 0
-        ul = "\x1b[%sm" * 3 % (2, 4, 33)
-        cols = ["\x1b[%sm" % n for n in range(91, 96)]
-        red, green, yellow, blue, pink = cols
-
-    r, g, y, b, p, w = red, green, yellow, blue, pink, white
-
-    @classmethod
-    def c(cls, colour, text):
-        """ Return coloured text. """
-        return getattr(cls, colour) + text + cls.w
 
 
 def setconfig(key, val):
@@ -4676,11 +4233,6 @@ def main():
                 sys.exit("Bad syntax")
 
         screen_update()
-
-def dbg(*args):
-    """Emit a debug message."""
-    # Uses xenc to deal with UnicodeEncodeError when writing to terminal
-    logging.debug(xenc(i) for i in args)
 
 g.helptext = [
     ("basic", "Basics", """
